[FSTORE-933] DBT Scheduling with External FG (#176)
* Tutorial for DBT with External FG
